requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15
django-environ>=0.11
dingtalk-stream>=0.21
dashscope>=1.19
httpx[socks]>=0.27
notion-client>=2.2.1
apscheduler>=3.10
pycryptodome>=3.20.0
beautifulsoup4>=4.12
lxml>=5.1
